chat dev branch

+++ CodeFreak1/Controllers/CompilerController.cs
@@ -96,7 +96,7 @@ namespace CodeFreak1.Controllers
 
                 }
                 Problem prob = probReops.getProblemById(Guid.Parse(code.problemId));
-                int maxScore = prob.MaxScore;
+                int maxScore = (int)prob.MaxScore;
                 if (flag == false)
                 {
                     submission.Status = "passed";
@@ -123,6 +123,7 @@ namespace CodeFreak1.Controllers
         }
         public Users getApplicationUser()
         {
+
             var identity = User.Identities.FirstOrDefault(s => s.Name.ToLower() == "user");
             var claims = identity.Claims;
             string id = null;
